code/requirements.txt
Flask==2.1.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
python-dotenv==0.20.0